requirements.txt
anyio==3.6.2
asyncio==3.4.3
certifi==2023.5.7
charset-normalizer==3.1.0
freeopcua==0.90.6
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
install==1.3.5
lxml==4.9.2
numpy==1.24.3
opcua==0.98.13
opencv-python==4.7.0.72
Pillow==9.5.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
pyzbar==0.1.9
requests==2.31.0
six==1.16.0
sniffio==1.3.0
urllib3==2.0.2
